Fixed error message of the Unregister File Extensions When Registry key's are not existing anymore do nothing, when they are remove the registry key's

+++ AnnoDesigner/MainWindow.xaml.cs
@@ -475,11 +475,17 @@ namespace AnnoDesigner
 
         private void MenuItemUnregisterExtensionClick(object sender, RoutedEventArgs e)
         {
-            // removes the registry entries            
-            Registry.CurrentUser.DeleteSubKeyTree(@"Software\Classes\anno_designer");
-            Registry.CurrentUser.DeleteSubKeyTree(@"Software\Classes\.ad");
+            var regCheckADFileExtensionPath = @"Software\Classes\anno_designer";
+            var localMachine = RegistryKey.OpenBaseKey(RegistryHive.CurrentUser, RegistryView.Default);
+            var regCheckADFileExtension = localMachine.OpenSubKey(regCheckADFileExtensionPath, false);
+            if (regCheckADFileExtension != null)
+            {
+                // removes the registry entries when exists          
+                Registry.CurrentUser.DeleteSubKeyTree(@"Software\Classes\anno_designer");
+                Registry.CurrentUser.DeleteSubKeyTree(@"Software\Classes\.ad");
 
-            showRegistrationMessageBox(isDeregistration: true);
+                showRegistrationMessageBox(isDeregistration: true);
+            }
         }
 
         private void showRegistrationMessageBox(bool isDeregistration)
